#358 complete migration of f09_filter to fo8_filter, f08_brick to f09_brick
