+++ src/f08_filter/examples/example_filters.py
@@ -7,7 +7,7 @@ from src.f04_gift.atom_config import (
     type_GroupID_str,
     type_RoadNode_str,
 )
-from src.f09_filter.filter import (
+from src.f08_filter.filter import (
     bridgeunit_shop,
     BridgeUnit,
     FilterUnit,
