+++ src/f08_filter/test/test_bridgeunit_dict.py
@@ -5,24 +5,16 @@ from src.f04_gift.atom_config import (
     type_RoadUnit_str,
     type_GroupID_str,
 )
-from src.f09_brick.pandas_tool import (
-    get_ordered_csv,
-    get_sorting_priority_column_headers as sorting_columns,
-)
-from src.f09_filter.filter import (
+from src.f08_filter.filter import (
     bridgeunit_shop,
     get_bridgeunit_from_dict,
     get_bridgeunit_from_json,
-    get_otx_to_inx_dt_columns,
-    get_explicit_label_columns,
-    create_otx_to_inx_dt,
-    create_explicit_label_dt,
 )
-from src.f09_filter.examples.filter_env import (
+from src.f08_filter.examples.filter_env import (
     env_dir_setup_cleanup,
     get_test_faces_dir,
 )
-from src.f09_filter.examples.example_filters import (
+from src.f08_filter.examples.example_filters import (
     get_casa_maison_filterunit_set_by_otx_to_inx,
     get_casa_maison_filterunit_set_by_explicit_label,
     get_casa_maison_road_otx_to_inx_dt,
@@ -154,77 +146,3 @@ def test_get_bridgeunit_from_json_ReturnsObj():
 
     # THEN
     assert x_bridgeunit == roadnode_bridgeunit
-
-
-def test_get_otx_to_inx_dt_columns_ReturnsObj():
-    # ESTABLISH / WHEN /THEN
-    assert get_otx_to_inx_dt_columns()
-    assert len(get_otx_to_inx_dt_columns()) == 7
-    static_list = [
-        "face_id",
-        "obj_class",
-        "otx_road_delimiter",
-        "inx_road_delimiter",
-        "unknown_word",
-        "otx_word",
-        "inx_word",
-    ]
-    assert get_otx_to_inx_dt_columns() == static_list
-    assert set(get_otx_to_inx_dt_columns()).issubset(set(sorting_columns()))
-
-
-def test_create_otx_to_inx_dt_ReturnsObj():
-    # ESTABLISH
-    casa_filterunit = get_casa_maison_filterunit_set_by_otx_to_inx()
-    casa_bridgeunit = casa_filterunit.get_bridgeunit(type_RoadUnit_str())
-
-    # WHEN
-    casa_dataframe = create_otx_to_inx_dt(casa_bridgeunit)
-    print(f"{casa_dataframe=}")
-
-    # THEN
-    assert list(casa_dataframe.columns) == get_otx_to_inx_dt_columns()
-    assert len(casa_dataframe) == 4
-    casa_csv = get_ordered_csv(casa_dataframe)
-    print(f"{casa_csv=}")
-    print(f"{get_ordered_csv(get_casa_maison_road_otx_to_inx_dt())=}")
-    assert casa_csv == get_ordered_csv(get_casa_maison_road_otx_to_inx_dt())
-
-
-def test_get_explicit_label_columns_ReturnsObj():
-    # ESTABLISH / WHEN /THEN
-    assert get_explicit_label_columns()
-    assert len(get_explicit_label_columns()) == 7
-    static_list = [
-        "face_id",
-        "obj_class",
-        "otx_road_delimiter",
-        "inx_road_delimiter",
-        "unknown_word",
-        "otx_label",
-        "inx_label",
-    ]
-    assert get_explicit_label_columns() == static_list
-    assert set(get_explicit_label_columns()).issubset(set(sorting_columns()))
-
-
-def test_create_explicit_label_dt_ReturnsObj():
-    # ESTABLISH
-    casa_filterunit = get_casa_maison_filterunit_set_by_explicit_label()
-    casa_bridgeunit = casa_filterunit.get_bridgeunit(type_RoadUnit_str())
-
-    # WHEN
-    casa_dataframe = create_explicit_label_dt(casa_bridgeunit)
-
-    # THEN
-    # print(f"{get_explicit_label_columns()=}")
-    # print(f"    {list(casa_dataframe.columns)=}")
-    # print("")
-    # print(f"{casa_dataframe=}")
-    assert list(casa_dataframe.columns) == get_explicit_label_columns()
-    assert len(casa_dataframe) == 3
-    casa_csv = get_ordered_csv(casa_dataframe)
-    ex_explicit_csv = get_ordered_csv(get_casa_maison_road_explicit_label_dt())
-    print(f"       {casa_csv=}")
-    print(f"{ex_explicit_csv=}")
-    assert casa_csv == ex_explicit_csv
