+++ src/f09_brick/dir_func.py
@@ -1,7 +1,7 @@
 from src.f00_instrument.file import get_dir_file_strs, open_file
 from src.f04_gift.atom_config import get_atom_args_obj_classs
 from src.f09_brick.pandas_tool import save_dataframe_to_csv, open_csv
-from src.f09_filter.filter import (
+from src.f08_filter.filter import (
     BridgeUnit,
     FilterUnit,
     filterable_atom_args,
